src/datamodules/blood_cells_datamodule.py
from typing import Any, Dict, Optional

from pytorch_lightning import LightningDataModule
from torch.utils.data import DataLoader, Dataset
from torchvision.transforms import transforms

from src.dataset.dataset import CellsDataset


class BloodCellsDataModule(LightningDataModule):
    """LightningDataModule for Blood cells dataset."""

    def __init__(
        self,
        train_data_dir,
        valid_data_dir,
        test_data_dir,
        batch_size: int = 64,
        num_workers: int = 0,
        pin_memory: bool = False,
    ):
        super().__init__()

        self.save_hyperparameters(logger=False)

        # data transformations
        self.train_transforms = transforms.Compose(
            [
                transforms.ToTensor(),
                transforms.Normalize((0.485, 0.456, 0.406), (0.229, 0.224, 0.225)),
                transforms.RandomRotation(90),
                transforms.RandomHorizontalFlip(),
            ]
        )
        self.test_transforms = transforms.Compose(
            [
                transforms.ToTensor(),
                transforms.Normalize((0.485, 0.456, 0.406), (0.229, 0.224, 0.225)),
            ]
        )
        self.data_train: Optional[Dataset] = None
        self.data_val: Optional[Dataset] = None
        self.data_test: Optional[Dataset] = None

    @property
    def num_classes(self):
        return 4

    def setup(self, stage: Optional[str] = None):
        """Load data."""

        self.data_train = CellsDataset(
            self.hparams.train_data_dir, transform=self.train_transforms
        )
        self.data_val = CellsDataset(self.hparams.valid_data_dir, self.test_transforms)
        self.data_test = CellsDataset(self.hparams.test_data_dir, self.test_transforms)

    def train_dataloader(self):

        return DataLoader(
            dataset=self.data_train,
            batch_size=self.hparams.batch_size,
            num_workers=self.hparams.num_workers,
            pin_memory=self.hparams.pin_memory,
            shuffle=True,
        )

    def val_dataloader(self):
        return DataLoader(
            dataset=self.data_val,
            batch_size=self.hparams.batch_size,
            num_workers=self.hparams.num_workers,
            pin_memory=self.hparams.pin_memory,
            shuffle=True,
        )

    def test_dataloader(self):
        return DataLoader(
            dataset=self.data_test,
            batch_size=self.hparams.batch_size,
            num_workers=self.hparams.num_workers,
            pin_memory=self.hparams.pin_memory,
            shuffle=False,
        )

    def predict_dataloader(self):
        return DataLoader(
            dataset=self.data_test,
            batch_size=self.hparams.batch_size,
            num_workers=self.hparams.num_workers,
            pin_memory=self.hparams.pin_memory,
            shuffle=False,
        )

    def teardown(self, stage: Optional[str] = None):
        """Clean up after fit or test."""
        pass

    def state_dict(self):
        """Extra things to save to checkpoint."""
        return {}

    def load_state_dict(self, state_dict: Dict[str, Any]):
        """Things to do when loading checkpoint."""
        pass


if __name__ == "__main__":
    import hydra
    import omegaconf
    import pyrootutils

    root = pyrootutils.setup_root(__file__, pythonpath=True)
    print(root)
    cfg = omegaconf.OmegaConf.load(root / "configs" / "datamodule" / "blood_cells.yaml")
    cfg.data_dir = str(root / "../cropped_cells")
    print(cfg)
    _ = hydra.utils.instantiate(cfg)
